+++ backend/internal/compilerServer/build_error.go
@@ -1,4 +1,4 @@
-package compile
+package compilerServer
 
 import (
 	"fmt"
