+++ backend/internal/compilerServer/compiler.go
@@ -1,4 +1,4 @@
-package compile
+package compilerServer
 
 import (
 	"bytes"
