+++ backend/internal/compilerServer/daemon.go
@@ -1,4 +1,4 @@
-package compile
+package compilerServer
 
 import (
 	"bytes"
